Bug fixes
When posting the tasks and updating the task, it will directly navigate to another route and if you see the URL after the task submission it won't show the /tasks navigation, it will show something else and then it will cause some errors when take any action to the tasks(Mark as completed/not completed). This bug fix will fix that error.